Refactor httpd conf, pulp and tests
