+++ src/requirements.txt
@@ -1,3 +1,3 @@
-obsah >= 1.9.0
+obsah >= 1.9.1
 requests
 requests-oauthlib
