+++ src/roles/pulp/defaults/main.yaml
@@ -28,7 +28,7 @@ pulp_mirror: false
 pulp_register_foreman_proxy: true
 pulp_rhsm_url: "https://{{ ansible_facts['fqdn'] }}/rhsm"
 
-pulp_default_import_paths: "{{ pulp_mirror | ternary(['/var/lib/pulp/sync_imports'], ['/var/lib/pulp/sync_imports', '/var/lib/pulp/imports']) }}"
+pulp_default_import_paths: "{{ pulp_mirror | ternary([], ['/var/lib/pulp/sync_imports', '/var/lib/pulp/imports']) }}"
 pulp_default_export_paths: "{{ pulp_mirror | ternary([], ['/var/lib/pulp/exports']) }}"
 pulp_import_paths: []
 pulp_export_paths: []
@@ -75,7 +75,6 @@ pulp_settings_other_env:
   PULP_SMART_PROXY_PULP_URL: "{{ pulp_pulp_url }}"
   PULP_SMART_PROXY_MIRROR: "{{ pulp_mirror }}"
   PULP_SMART_PROXY_RHSM_URL: "{{ pulp_rhsm_url }}"
-  PULP_SMART_PROXY_AUTH_METHODS: ['client_certificate']
   PULP_API_WORKERS: "{{ pulp_api_service_worker_count }}"
   PULP_CONTENT_WORKERS: "{{ pulp_content_service_worker_count }}"
   PULP_TOKEN_AUTH_DISABLED: "true"
