+++ src/roles/pulp/tasks/main.yaml
@@ -25,6 +25,7 @@
     state: directory
     mode: "0755"
   loop: "{{ pulp_default_import_paths + pulp_import_paths }}"
+  when: not pulp_mirror
 
 - name: Create export directories
   ansible.builtin.file:
