+++ tests/foreman_api_test.py
@@ -7,61 +7,6 @@ def test_foreman_organization(organization):
     assert organization
 
 
-def test_foreman_product(product):
-    assert product
-
-
-@pytest.mark.feature('katello')
-def test_foreman_yum_repository(yum_repository, foremanapi, local_request):
-    assert yum_repository
-    foremanapi.resource_action('repositories', 'sync', {'id': yum_repository['id']})
-    repo_url = yum_repository['full_path']
-    assert local_request.get(f'{repo_url}/repodata/repomd.xml', verify=False)
-    assert local_request.get(f'{repo_url}/Packages/b/bear-4.1-1.noarch.rpm', verify=False)
-
-
-@pytest.mark.feature('katello')
-def test_foreman_file_repository(file_repository, foremanapi, local_request):
-    assert file_repository
-    foremanapi.resource_action('repositories', 'sync', {'id': file_repository['id']})
-    repo_url = file_repository['full_path']
-    assert local_request.get(f'{repo_url}/1.iso', verify=False)
-
-
-@pytest.mark.feature('katello')
-def test_foreman_container_repository(container_repository, foremanapi):
-    assert container_repository
-    foremanapi.resource_action('repositories', 'sync', {'id': container_repository['id']})
-
-
-@pytest.mark.feature('katello')
-def test_foreman_lifecycle_environment(lifecycle_environment):
-    assert lifecycle_environment
-
-
-@pytest.mark.feature('katello')
-def test_foreman_content_view(content_view, yum_repository, foremanapi):
-    assert content_view
-    foremanapi.update('content_views', {'id': content_view['id'], 'repository_ids': [yum_repository['id']]})
-    foremanapi.resource_action('content_views', 'publish', {'id': content_view['id']})
-    # do something with the published view
-    versions = foremanapi.list('content_view_versions', params={'content_view_id': content_view['id']})
-    for version in versions:
-        current_environment_ids = {environment['id'] for environment in version['environments']}
-        for environment_id in current_environment_ids:
-            foremanapi.resource_action('content_views', 'remove_from_environment', params={'id': content_view['id'], 'environment_id': environment_id})
-        foremanapi.delete('content_view_versions', version)
-
-
-@pytest.mark.feature('katello')
-def test_foreman_manifest(organization, foremanapi, fixture_dir):
-    manifest_path = fixture_dir / 'manifest.zip'
-    with open(manifest_path, 'rb') as manifest_file:
-        files = {'content': (str(manifest_path), manifest_file, 'application/zip')}
-        params = {'organization_id': organization['id']}
-        foremanapi.resource_action('subscriptions', 'upload', params, files=files)
-
-
 def test_foreman_initial_organization(foremanapi):
     assert foremanapi.list('organizations', search='name="Foreman CI"')
